// frontend/src/app/model/hotel.ts
export class Hotel {
    id: string;
    name: string;
    address: string;
    city: string;
    country: string;
    email: string;
    zipCode: string;
    description: string;
    insertUserId: string;
    updateUserId: string;
    insertDateTime: Date;
    updateDateTime: Date;
    status: number;
    sortOrder: number;
    averageRating: number;
}